Create prototype_1.ino
Code used to make a proof of concept for the sensors to be included in the gloves.

The potentiometer and flex sensor uses a voltage divider setup, with code to calibrate their range of possible values in practice.

To read the IMU, the mpu6050 library by ElectronicCats is used. The code to read yaw, pitch and roll largely follows that of the example.